Refactor ProjectData
