+++ src/app/pages/ProjectDetails.tsx
@@ -1,16 +1,6 @@
 import React from "react";
 import Header from "../components/Header";
-
-interface ProjectData {
-    title: string;
-    year: number;
-    type: string;
-    tools: Array<string>;
-    accomplishments: Array<string>;
-    about: string;
-    images: Array<string> | null; // Optional array of image URLs for the image gallery
-    youtubeLink: string | null; // Optional link to a YouTube video for the project
-}
+import { ProjectData } from "../../types/ProjectData";
 
 interface Project {
     project: ProjectData
@@ -56,4 +46,3 @@ const ProjectDetails: React.FC<Project> = ({ project }) => {
 }
 
 export { ProjectDetails };
-export type { ProjectData };
