+++ src/projects/space_miner.ts
@@ -1,4 +1,4 @@
-import { ProjectData } from "../app/pages/ProjectDetails"
+import { ProjectData } from "../types/ProjectData";
 
 const spaceMiner: ProjectData = {
     title: "Space Miner",
